Use revision to detect unpublished changes state (+ .js -> .ts)
